metadata_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg[binary]>=3.1
pydantic
python-dotenv